* Implemented the Archive command.
I needed to modify associated classes that proxy to Archive to make it work by removing code that assumed a fluent API. I pondered some ways to get the fluent API working in the comments and will branch the code to experiment further with it.

This code needs a good clean up of comments and some refactoring to ensure labels are consistent across classes and more checking of assumptions.